pyproject.toml
[project]
name = "paddydabaddie"
version = "0.1.0"
description = "Ask Hume — David Hume's philosophy, Paddy Pimblett's mouth"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "anthropic>=0.40.0",
    "openai>=1.57.0",
    "chromadb>=0.6.0",
    "elevenlabs>=1.10.0",
    "sentence-transformers>=3.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "tiktoken>=0.8.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
